backend/requirements.txt
# Swish Vision Backend Requirements

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ML & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
supervision>=0.16.0

# Roboflow Inference
inference>=0.9.0
roboflow>=1.1.0

# SAM2 (install separately)
# pip install git+https://github.com/facebookresearch/sam2.git

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
